+++ mangadex_library/lib/models/common/future_updates.dart
@@ -1,0 +1,4 @@
+enum FutureUpdates {
+  enable,
+  disable,
+}