+++ mangadex_library/lib/models/common/language_codes.dart
@@ -1,0 +1,11 @@
+enum LanguageCodes {
+  en, // English
+  zh, // Simplified Chinese
+  zh_hk, // Traditional Chinese
+  pt_br, // Brazillian Portugese
+  es, // Castilian Spanish
+  es_la, // Latin American Spanish
+  ja_ro, // Romanized Japanese
+  ko_ro, // Romanized Korean
+  zh_ro, // Romanized Chinese
+}